+++ ChangeLog.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add support for contract schemas containing the version.
+- Add support for the contract schema types, ULeb128, ILeb128, ByteList and ByteArray.
+
 ## 4.0.3
 
 - Fix calculation of payload size of a configure baker transaction in